Move the code to display and edit the sync settings
Similar to
https://github.com/e-mission/e-mission-data-collection/pull/153,
modify the sync plugin to include code to display and edit its settings.

The app does not need to include the settings (although it would probably be
good if it could), but it does ensure that any changes to the config format can
be made in one place.